Remove old menus
