+++ components/Menu/Backdrop.js
@@ -1,19 +1,29 @@
 // @flow
 /**
  * Part of GDL gdl-frontend.
- * Copyright (C) 2017 GDL
+ * Copyright (C) 2018 GDL
  *
  * See LICENSE
  */
+import React from 'react';
 import styled from 'react-emotion';
 
-const Backdrop = styled.div`
-  position: fixed;
+/**
+ * Backdrop/Blanket component to use behind modals etc.
+ */
+
+type Props = {
+  onClick: (event: SyntheticMouseEvent<any>) => void
+};
+
+const StyledBackdrop = styled('div')`
   bottom: 0;
   left: 0;
+  position: fixed;
   right: 0;
   top: 0;
-  z-index: 100;
 `;
+
+const Backdrop = (props: Props) => <StyledBackdrop {...props} />;
 
 export default Backdrop;
