--- components/Menu/Container.js
@@ -1,17 +1,0 @@
-// @flow
-/**
- * Part of GDL gdl-frontend.
- * Copyright (C) 2017 GDL
- *
- * See LICENSE
- */
-
-import styled from 'react-emotion';
-import Container from '../Container';
-
-export default styled(Container)`
-  padding-left: 0;
-  padding-right: 0;
-  display: flex;
-  height: 100%;
-`;
