--- components/Menu/IconButton.js
@@ -1,9 +1,0 @@
-import styled from 'react-emotion';
-
-const IconButton = styled('button')`
-  background: transparent;
-  border: none;
-  color: inherit;
-`;
-
-export default IconButton;
