+++ components/Menu/MenuItem.js
@@ -1,35 +1,86 @@
 // @flow
 /**
  * Part of GDL gdl-frontend.
- * Copyright (C) 2017 GDL
+ * Copyright (C) 2018 GDL
  *
  * See LICENSE
  */
-import styled from 'react-emotion';
-import theme from '../../style/theme';
+import React, { type Node } from 'react';
+import { MdCheck, MdKeyboardArrowRight } from 'react-icons/lib/md';
+import { Trans } from '@lingui/react';
+import { cx } from 'react-emotion';
 
-const MenuItem = styled.a`
-  font-size: 14px;
-  display: flex;
-  align-items: center;
+import SrOnly from '../SrOnly';
+import { Item, ItemIcon, itemActionStyle } from './styled/Content';
 
-  padding: 7px 15px;
-  min-height: 38px;
-  width: 100%;
-  transition: background-color 0.2s ease;
+type Props = {
+  isSelected?: boolean,
+  showKeyLine?: boolean,
+  href?: string,
+  onClick?: (event: SyntheticMouseEvent<any>) => void,
+  onCustomClick?: (event: SyntheticMouseEvent<any>) => void,
+  hasNestedMenu?: boolean,
+  children: Node
+};
 
-  &:hover {
-    background-color: ${theme.colors.highlight};
+const ItemLink = Item.withComponent('a');
+
+class MenuItem extends React.Component<Props> {
+  /**
+   * Small trick to handle Next.js swallowing up any onClicks on Link components
+   * See https://github.com/zeit/next.js/issues/1490#issuecomment-290724312
+   */
+  handleClick = (event: SyntheticMouseEvent<any>) => {
+    this.props.onClick && this.props.onClick(event);
+
+    this.props.onCustomClick && this.props.onCustomClick(event);
+  };
+
+  handleKeyDown = (event: SyntheticKeyboardEvent<any>) => {
+    if (this.props.onClick && event.key === 'Enter') {
+      console.log(event);
+      this.props.onClick(event);
+    }
+  };
+
+  render() {
+    const {
+      isSelected,
+      hasNestedMenu,
+      showKeyLine,
+      children,
+      onClick,
+      href
+    } = this.props;
+    const ItemComponent = href ? ItemLink : Item;
+
+    return (
+      <ItemComponent
+        className={cx({ [itemActionStyle]: href || onClick })}
+        onClick={this.handleClick}
+        onKeyDown={this.handleKeyDown}
+        href={href}
+        isSelected={isSelected}
+        showKeyLine={showKeyLine}
+        tabIndex={href ? null : onClick ? '0' : null}
+      >
+        {isSelected && (
+          <ItemIcon>
+            <MdCheck aria-hidden />
+            <SrOnly>
+              <Trans>Selected:</Trans>
+            </SrOnly>
+          </ItemIcon>
+        )}
+        {children}
+        {hasNestedMenu && (
+          <ItemIcon aria-hidden style={{ right: '10px', left: 'unset' }}>
+            <MdKeyboardArrowRight />
+          </ItemIcon>
+        )}
+      </ItemComponent>
+    );
   }
-
-  &:active {
-    color: ${theme.colors.dark};
-  }
-
-  ${p =>
-    p.thinBorder &&
-    `border-bottom: 1px solid ${theme.colors.grayLight};`} ${p =>
-      p.thickBorder && `border-bottom: 2px solid ${theme.colors.grayLight};`};
-`;
+}
 
 export default MenuItem;
