--- components/Menuu/index.js
@@ -1,10 +1,0 @@
-// @flow
-/**
- * Part of GDL gdl-frontend.
- * Copyright (C) 2018 GDL
- *
- * See LICENSE
- */
-
-export { default } from './Menu';
-export { default as MenuItem } from './MenuItem';
