--- components/Sidemenu/index.js
@@ -1,11 +1,0 @@
-// @flow
-/**
- * Part of GDL gdl-frontend.
- * Copyright (C) 2017 GDL
- *
- * See LICENSE
- */
-
-import Sidemenu from './Sidemenu';
-
-export default Sidemenu;
