+++ package.json
@@ -50,7 +50,6 @@
     "compression": "1.7.2",
     "cookie-parser": "1.4.3",
     "domurl": "2.1.7",
-    "downshift": "1.28.0",
     "emotion": "9.0.1",
     "emotion-server": "9.0.1",
     "emotion-theming": "9.0.0",
